+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [breaking] - Drop support of `Nya` (Chewa)
 * [breaking] - Rename Yiddish: `Ydd` -> `Yid`
 * [breaking] - Rename Azerbaijani: `Azj` -> `Aze`
+* Drop support of Cyrillic Azerbaijani and Turkmen
 * Add `Script::values()` to iterate over all scripts.
 * Add `Lang::values()` to iterate over all languages.
 * Add integration with `enum-map`